feat(ratelimit): make tokens regenerate based on refill rate

+++ gateway/src/server/upstream/models.rs
@@ -27,7 +27,7 @@ pub struct UpstreamServer {
     pub pool: ConnectionPool,
     pub req_auth: bool,
     // TODO: should not be in upstream but testing it for now here.
-    pub token_weight: usize,
+    pub token_weight: f64,
 }
 
 pub struct ConnectionPool {
@@ -44,7 +44,7 @@ impl UpstreamServer {
         max_connections: usize,
         protocol: Protocol,
         req_auth: bool,
-        token_weight: usize,
+        token_weight: f64,
     ) -> Self {
         UpstreamServer {
             active_connctions: AtomicUsize::new(0),
